+++ server/src/klite/HttpExchange.kt
@@ -35,10 +35,12 @@ open class HttpExchange(
   val path: String get() = original.requestURI.path
   lateinit var pathParams: Params internal set
   fun path(param: String): String? = pathParams[param]
+  inline fun <reified T: Any> path(param: String): T? = path(param)?.let { Converter.from<T>(it) }
 
   val query: String get() = original.requestURI.query?.let { "?$it" } ?: ""
   val queryParams: Params by lazy { original.requestURI.queryParams }
   fun query(param: String): String? = queryParams[param]
+  inline fun <reified T: Any> query(param: String): T? = query(param)?.let { Converter.from<T>(it) }
 
   val fullUrl get() = fullUrl(original.requestURI.toString())
   fun fullUrl(suffix: String): URI = URI("$protocol://$host$suffix")